fix: some dependency cannot be resolved with value updating
The CPU uses the data sent from Load & Store Buffer and Reservation Station
to solve the dependency of some value. However, the data have one-clock delay,
so if the data is updated when another instruction that depends on it is at
instruction issuing stage, the value of that instruction will neither update
when issuing nor when adding to Load & Store Buffer or Reservation Station.

This commit fixes that problem by updating the data at the adding stage.