fix: bugs in Load & Store Buffer on wrong branch prediction
There is a register called `processing' in Load & Store Buffer to indicate
whether there is a ongoing work in DCache module. According to the document,
when there is a wrong branch prediction, DCache will abort memory read.

Therefore, Load & Store Buffer should set the `processing' register to be 0,
if there is a memory read work ,or if the data write work is done right when
the `clear' signal that indicates a wrong branch prediction is high.