fix: a bug about time delay in Load & Store Buffer
In the previous implementation, the Load & Store Buffer module will change
the update index in Reorder Buffer as soon as a new line is sent to the
data cache. However, if that happens when the data have just finished
loading a value, the data valid may still be high considering the time delay.
This will cause other modules update the data that needs the data that is
set to data cache with wrong value.

This commit fixes this problem by assign the value after a clock time.